+++ projects/core-ui/core-ui-packagr/index.js
@@ -78,6 +78,7 @@ function copyFiles(context) {
             { source: '../src/lib/global-alert/**', dest: '../../../dist/core-ui/schematics/ng-add/files/src/app/components/global-alert/' },
             { source: '../src/lib/global-progress/**', dest: '../../../dist/core-ui/schematics/ng-add/files/src/app/components/global-progress/' },
             { source: '../src/lib/header/**', dest: '../../../dist/core-ui/schematics/ng-add/files/src/app/components/header/' },
+            { source: '../src/lib/icons/**', dest: '../../../dist/core-ui/schematics/ng-add/files/src/app/components/icons/' },
             { source: '../schematics/migrations/**', dest: '../../../dist/core-ui/schematics/migrations/' }
         ];
         const afilePaths = filePaths.map(item => ({ source: (0, path_1.join)(__dirname, item.source), dest: (0, path_1.join)(__dirname, item.dest) }));