+++ projects/core-ui/src/public-api.ts
@@ -7,6 +7,7 @@ export * from './lib/global-progress/progress.service';
 export * from './lib/global-alert/global-alert.component';
 export * from './lib/global-alert/alert.service';
 export * from './lib/header/header.component';
+export * from './lib/icons/ukis';
 export * from './lib/dynamic-component/dynamic-component.component';
 
 export * from './lib/core-ui.module';